requirements.txt
# FastAPI Framework
fastapi==0.109.0
flask
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# Firebase Admin SDK
firebase-admin==6.4.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP Requests
httpx==0.26.0
requests==2.31.0

# Email Support (for password reset)
fastapi-mail==1.4.1
aiosmtplib==3.0.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Image Handling (for future OCR implementation)
Pillow==10.2.0

# Date & Time Handling
python-dateutil==2.8.2

# CORS Middleware
fastapi-cors==0.0.6

# JSON & Data Processing
orjson==3.9.12

# Email Support (Optional - for notifications)
fastapi-mail==1.4.1

# Background Tasks (Optional)
celery==5.3.4
redis==5.0.1

# Logging
loguru==0.7.2

# Currency Conversion
forex-python==1.8

# PDF Generation (Optional - for expense reports)
reportlab==4.0.9
PyPDF2==3.0.1

# Excel Export (Optional - for expense reports)
openpyxl==3.1.2
xlsxwriter==3.1.9

# OCR
pip install pytesseract==0.3.10
Pillow==10.2.0